+++ packages/plugins/Draw/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## unpublished
 
+- Feature: Add new configuration parameter `measureOptions` to allow users to select a measurement mode when drawing feature. This way, a length / area in the selected unit is added to the drawn feature.
 - Chore: Add `@polar/core` as a dependency as the component `RadioCard.vue` has been moved from this package to `@polar/core`.
 
 ## 2.0.0