+++ packages/plugins/Draw/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - Feature: Add new configuration parameter `measureOptions` to allow users to select a measurement mode when drawing a feature. This way, a length / area in the selected unit is added to the drawn feature.
 - Fix: Update initial value of `drawMode` to a selectable value if the default `Point` is not a drawable option.
-- Fix: Adjust types `selectableDrawModes` and `selectableModes` to correctly reflect that they represent an object.
+- Fix: Adjust type `DrawGetters` regarding its keys `selectableDrawModes` and `selectableModes` to correctly reflect that they represent objects.
 - Chore: Add `@polar/core` as a dependency as the component `RadioCard.vue` has been moved from this package to `@polar/core`.
 
 ## 2.0.0