+++ packages/plugins/Draw/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## unpublished
 
+- Breaking: Upgrade peerDependency `ol` from `^9.2.4` to `^10.3.1`.
 - Feature: Add new configuration parameter `measureOptions` to allow users to select a measurement mode when drawing a feature. This way, a length / area in the selected unit is added to the drawn feature.
 - Fix: Update initial value of `drawMode` to a selectable value if the default `Point` is not a drawable option.
 - Fix: Adjust type `DrawGetters` regarding its keys `selectableDrawModes` and `selectableModes` to correctly reflect that they represent objects.